moved Dict type to qgiv-server
