+++ libs/types/index.ts
@@ -1,2 +1,1 @@
-export * from './data-structures';
 export * from './http';
